requirements.txt
git+https://gitlab.com/nekokatt/hikari@feature/http_adapter
pyyaml==5.3
asyncpg==0.20.0
pydantic==1.3
